Refactor SchemaAwareLogBuilder into a package.
